avformat/wavdec: Fix overflow of intermediate in block_align check
Fixes: signed integer overflow: -251517880 * 32 cannot be represented in type 'int'
Fixes: 385224934/clusterfuzz-testcase-minimized-ffmpeg_dem_W64_fuzzer-4909298151915520

Found-by: continuous fuzzing process https://github.com/google/oss-fuzz/tree/master/projects/ffmpeg

+++ libavformat/wavdec.c
@@ -912,10 +912,10 @@ static int w64_read_header(AVFormatContext *s)
             if (st->codecpar->block_align &&
                 st->codecpar->ch_layout.nb_channels < FF_SANE_NB_CHANNELS &&
                 st->codecpar->bits_per_coded_sample < 128) {
-                int block_align = st->codecpar->block_align;
+                int64_t block_align = st->codecpar->block_align;
 
                 block_align = FFMAX(block_align,
-                                    ((st->codecpar->bits_per_coded_sample + 7) / 8) *
+                                    ((st->codecpar->bits_per_coded_sample + 7LL) / 8) *
                                     st->codecpar->ch_layout.nb_channels);
                 if (block_align > st->codecpar->block_align) {
                     av_log(s, AV_LOG_WARNING, "invalid block_align: %d, broken file.\n",
